Create ModelTransformerVersion2.cs with non-implemented parameters rule
