+++ src/Markdown.MAML/Markdown.MAML.csproj
@@ -79,6 +79,7 @@
     <Compile Include="Transformer\IModelTransformer.cs" />
     <Compile Include="Transformer\ModelTransformerBase.cs" />
     <Compile Include="Transformer\ModelTransformerVersion1.cs" />
+    <Compile Include="Transformer\ModelTransformerVersion2.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
